Refactor NSD4 UT for using upstream eventlet
This patch refactor NSD4 UT by mocking out instead of using eventlet
for fake NSD4 backend.

This patch also remove pinned eventlet version and use upper constraint
of eventlet for UT.

Closes-Bug: 1686601

+++ designate/tests/test_backend/test_nsd4.py
@@ -14,62 +14,16 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import os
 import socket
 import ssl
 
 import eventlet
-import fixtures
-from mock import MagicMock
+import mock
 
 from designate import exceptions
 from designate import objects
 from designate.tests.test_backend import BackendTestCase
-from designate.tests import resources
 from designate.backend import impl_nsd4
-
-
-class NSD4ServerStub(object):
-    recved_command = None
-    response = 'ok'
-    keyfile = os.path.join(resources.path, 'ssl', 'nsd_server.key')
-    certfile = os.path.join(resources.path, 'ssl', 'nsd_server.pem')
-
-    def handle(self, client_sock, client_addr):
-        stream = client_sock.makefile()
-        self.recved_command = stream.readline()
-        stream.write(self.response)
-        stream.flush()
-
-    def start(self):
-        self.port = 1025
-        while True:
-            try:
-                eventlet.spawn_n(eventlet.serve,
-                                 eventlet.wrap_ssl(
-                                     eventlet.listen(('127.0.0.1', self.port)),
-                                     keyfile=self.keyfile,
-                                     certfile=self.certfile,
-                                     server_side=True),
-                                 self.handle)
-                break
-            except socket.error:
-                self.port = self.port + 1
-
-    def stop(self):
-        eventlet.StopServe()
-
-
-class NSD4Fixture(fixtures.Fixture):
-    def setUp(self):
-        super(NSD4Fixture, self).setUp()
-        self.server = NSD4ServerStub()
-        self.server.start()
-
-        self.addCleanup(self.tearDown)
-
-    def tearDown(self):
-        self.server.stop()
 
 
 # NOTE: We'll only test the specifics to the nsd4 backend here.
@@ -78,51 +32,72 @@ class NSD4BackendTestCase(BackendTestCase):
     def setUp(self):
         super(NSD4BackendTestCase, self).setUp()
 
-        self.server_fixture = NSD4Fixture()
-        self.useFixture(self.server_fixture)
-
-        keyfile = os.path.join(resources.path, 'ssl', 'nsd_control.key')
-        certfile = os.path.join(resources.path, 'ssl', 'nsd_control.pem')
-
+        # NOTE(hieulq): we mock out NSD4 back-end with random port
+        self.port = 6969
         self.target = objects.PoolTarget.from_dict({
             'id': '4588652b-50e7-46b9-b688-a9bad40a873e',
             'type': 'nsd4',
             'masters': [{'host': '192.0.2.1', 'port': 53},
                         {'host': '192.0.2.2', 'port': 35}],
             'options': [
-                {'key': 'keyfile', 'value': keyfile},
-                {'key': 'certfile', 'value': certfile},
+                {'key': 'keyfile', 'value': mock.sentinel.key},
+                {'key': 'certfile', 'value': mock.sentinel.cert},
                 {'key': 'pattern', 'value': 'test-pattern'},
-                {'key': 'port', 'value': self.server_fixture.server.port}
+                {'key': 'port', 'value': self.port}
             ],
         })
-
         self.backend = impl_nsd4.NSD4Backend(self.target)
 
-    def test_create_zone(self):
+    @mock.patch.object(eventlet, 'connect')
+    @mock.patch.object(eventlet, 'wrap_ssl')
+    def _test_command(self, mock_ssl, mock_connect, command_context):
+        sock = mock.MagicMock()
+        stream = mock.MagicMock()
+        mock_connect.return_value = mock.sentinel.client
+        mock_ssl.return_value = sock
+        sock.makefile.return_value = stream
+        if command_context is 'create_fail':
+            stream.read.return_value = 'goat'
+        else:
+            stream.read.return_value = 'ok'
+
         context = self.get_context()
         zone = self.get_zone_fixture()
-        self.backend.create_zone(context, zone)
-        command = 'NSDCT1 addzone %s test-pattern\n' % zone['name']
-        self.assertEqual(command, self.server_fixture.server.recved_command)
+
+        if command_context is 'create':
+            self.backend.create_zone(context, zone)
+            command = 'NSDCT1 addzone %s test-pattern\n' % zone['name']
+        elif command_context is 'delete':
+            self.backend.delete_zone(context, zone)
+            command = 'NSDCT1 delzone %s\n' % zone['name']
+        elif command_context is 'create_fail':
+            self.assertRaises(exceptions.Backend,
+                              self.backend.create_zone,
+                              context, zone)
+            command = 'NSDCT1 addzone %s test-pattern\n' % zone['name']
+
+        stream.write.assert_called_once_with(command)
+        mock_ssl.assert_called_once_with(mock.sentinel.client,
+                                         certfile=mock.sentinel.cert,
+                                         keyfile=mock.sentinel.key)
+        mock_connect.assert_called_once_with(('127.0.0.1', self.port))
+        sock.makefile.assert_called_once_with()
+        sock.close.assert_called_once_with()
+        stream.close.assert_called_once_with()
+        stream.flush.assert_called_once_with()
+        stream.read.assert_called_once_with()
+
+    def test_create_zone(self):
+        self._test_command(command_context='create')
 
     def test_delete_zone(self):
-        context = self.get_context()
-        zone = self.get_zone_fixture()
-        self.backend.delete_zone(context, zone)
-        command = 'NSDCT1 delzone %s\n' % zone['name']
-        self.assertEqual(command, self.server_fixture.server.recved_command)
+        self._test_command(command_context='delete')
 
     def test_server_not_ok(self):
-        self.server_fixture.server.response = 'goat'
-        context = self.get_context()
-        zone = self.get_zone_fixture()
-        self.assertRaises(exceptions.Backend,
-                          self.backend.create_zone,
-                          context, zone)
+        self._test_command(command_context='create_fail')
 
     def test_ssl_error(self):
-        self.backend._command = MagicMock(side_effect=ssl.SSLError)
+        self.backend._command = mock.MagicMock(side_effect=ssl.SSLError)
         context = self.get_context()
         zone = self.get_zone_fixture()
         self.assertRaises(exceptions.Backend,
@@ -130,7 +105,7 @@ class NSD4BackendTestCase(BackendTestCase):
                           context, zone)
 
     def test_socket_error(self):
-        self.backend._command = MagicMock(side_effect=socket.error)
+        self.backend._command = mock.MagicMock(side_effect=socket.error)
         context = self.get_context()
         zone = self.get_zone_fixture()
         self.assertRaises(exceptions.Backend,
